requirements.txt
discord.py
python-docx